implement two factor authentication and api

+++ routes/api.php
@@ -23,26 +23,26 @@ use App\Http\Controllers\Api\V1\ELearning\UserSavedLearningPathController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
-Route::middleware(['auth:sanctum'])->get('/user', function (Request $request) {
+Route::middleware(['auth:sanctum','auth.two_factor'])->get('/user', function (Request $request) {
     return $request->user()->load('permissions:name');
 });
 
 Route::get('/user/abilities', AbilityController::class);
 
 // ========== Courses ==========
-Route::get('/courses/interests', GetCoursesBasedOnUserInterest::class)->middleware('auth:sanctum');
-Route::get('/courses/interests/{tag_id}/beginner', GetInterestTagBeginnerCourseController::class)->middleware('auth:sanctum');
-Route::get('/courses/recommended-for-user-interest', GetRecommendedCourseForUserInterest::class)->middleware('auth:sanctum');
-Route::get('/learning-paths/recommended', GetRecommendedLearningPathController::class)->middleware('auth:sanctum');
+Route::get('/courses/interests', GetCoursesBasedOnUserInterest::class)->middleware(['auth:sanctum','auth.two_factor']);
+Route::get('/courses/interests/{tag_id}/beginner', GetInterestTagBeginnerCourseController::class)->middleware(['auth:sanctum','auth.two_factor']);
+Route::get('/courses/recommended-for-user-interest', GetRecommendedCourseForUserInterest::class)->middleware(['auth:sanctum','auth.two_factor']);
+Route::get('/learning-paths/recommended', GetRecommendedLearningPathController::class)->middleware(['auth:sanctum','auth.two_factor']);
 Route::get('/courses/new-and-popular', GetNewAndPopularCourseController::class);
 Route::get('/search', GetSearchResultController::class);
 
 // ========== For User Interest Tag ==========
 Route::get('/skill-tags', GetSkillTagController::class);
 
-Route::get('/followed-tags', [UserInterestTagController::class, 'index'])->middleware('auth:sanctum');
-Route::post('/follow-tag', [UserInterestTagController::class, 'followTag'])->middleware('auth:sanctum');
-Route::post('/unfollow-tag', [UserInterestTagController::class, 'unFollowTag'])->middleware('auth:sanctum');
+Route::get('/followed-tags', [UserInterestTagController::class, 'index'])->middleware(['auth:sanctum','auth.two_factor']);
+Route::post('/follow-tag', [UserInterestTagController::class, 'followTag'])->middleware(['auth:sanctum','auth.two_factor']);
+Route::post('/unfollow-tag', [UserInterestTagController::class, 'unFollowTag'])->middleware(['auth:sanctum','auth.two_factor']);
 
 // ========== For Blog Page ==========
 Route::get('/content/resources', GetResourcesForBlogPageController::class);
@@ -65,14 +65,14 @@ Route::get('/learning-paths', [LearningPathController::class, 'index']);
 Route::get('/learning-paths/{learning_path}', [LearningPathController::class, 'show']);
 
 // ========== For Saved List ==========
-Route::post('/courses/{course}/save', [UserSavedCourseController::class, 'saveCourse'])->middleware('auth:sanctum');
-Route::delete('/courses/{course}/remove', [UserSavedCourseController::class, 'removeSavedCourse'])->middleware('auth:sanctum');
-Route::post('/learning-paths/{learning_path}/save', [UserSavedLearningPathController::class, 'saveLearningPath'])->middleware('auth:sanctum');
-Route::delete('/learning-paths/{learning_path}/remove', [UserSavedLearningPathController::class, 'removeSavedLearningPath'])->middleware('auth:sanctum');
+Route::post('/courses/{course}/save', [UserSavedCourseController::class, 'saveCourse'])->middleware(['auth:sanctum','auth.two_factor']);
+Route::delete('/courses/{course}/remove', [UserSavedCourseController::class, 'removeSavedCourse'])->middleware(['auth:sanctum','auth.two_factor']);
+Route::post('/learning-paths/{learning_path}/save', [UserSavedLearningPathController::class, 'saveLearningPath'])->middleware(['auth:sanctum','auth.two_factor']);
+Route::delete('/learning-paths/{learning_path}/remove', [UserSavedLearningPathController::class, 'removeSavedLearningPath'])->middleware(['auth:sanctum','auth.two_factor']);
 
 // ========== Subscription Plan ==========
-Route::get('/user/stripe-intent', [StripeSubscriptionController::class,"getUserSetupIntent"])->middleware('auth');
-Route::post('/subscribe', [StripeSubscriptionController::class,"subscribe"])->middleware('auth');
+Route::get('/user/stripe-intent', [StripeSubscriptionController::class,"getUserSetupIntent"])->middleware(['auth:sanctum','auth.two_factor']);
+Route::post('/subscribe', [StripeSubscriptionController::class,"subscribe"])->middleware(['auth:sanctum','auth.two_factor']);
 
 require __DIR__.'/admin.php';
 require __DIR__.'/instructor.php';
